+++ AgingTestSystem/Properties/Resources.Designer.cs
@@ -60,24 +60,5 @@ namespace AgingTestSystem.Properties {
             }
         }
         
-        /// <summary>
-        ///   查找 System.Drawing.Bitmap 类型的本地化资源。
-        /// </summary>
-        internal static System.Drawing.Bitmap 华际光电_1__1_ {
-            get {
-                object obj = ResourceManager.GetObject("华际光电(1)(1)", resourceCulture);
-                return ((System.Drawing.Bitmap)(obj));
-            }
-        }
-        
-        /// <summary>
-        ///   查找 System.Drawing.Bitmap 类型的本地化资源。
-        /// </summary>
-        internal static System.Drawing.Bitmap 华际光电_1__1__1_ {
-            get {
-                object obj = ResourceManager.GetObject("华际光电(1)(1)(1)", resourceCulture);
-                return ((System.Drawing.Bitmap)(obj));
-            }
-        }
     }
 }